// fe/src/lib/api/generated/model/privacyActivityCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type PrivacyActivityCategory = typeof PrivacyActivityCategory[keyof typeof PrivacyActivityCategory];


export const PrivacyActivityCategory = {
  DATA_RELEASE: 'DATA_RELEASE',
  ACCESS_CONTROL: 'ACCESS_CONTROL',
  ACCOUNT_SECURITY: 'ACCOUNT_SECURITY',
  PRIVACY_GOVERNANCE: 'PRIVACY_GOVERNANCE',
} as const;