// fe/src/lib/api/generated/model/reviewTypeValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type ReviewTypeValue = typeof ReviewTypeValue[keyof typeof ReviewTypeValue];


export const ReviewTypeValue = {
  PRIVACY_REVIEW: 'PRIVACY_REVIEW',
  PIA: 'PIA',
} as const;